+++ Cargo.toml
@@ -66,6 +66,9 @@ ctrlc = { version = "3.2.3", features = ["termination"] }
 clap = { version = "3.2.18", features = ["derive", "env"] }
 dirs = "4.0.0"
 
+# Specular-specific
+enum_dispatch = "0.3.9"
+
 [features]
 default = ["test-utils"]
 test-utils = []
